+++ include/redis_cpp/detail/sync/redis_set.hpp
@@ -39,7 +39,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -57,7 +57,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -75,7 +75,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -92,7 +92,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd, values);
+        return (int32_t)get_integer_result(cmd, values);
     }
 
     /**
@@ -108,7 +108,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
