+++ include/redis_cpp/detail/sync/redis_zset.hpp
@@ -42,7 +42,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -66,7 +66,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -82,7 +82,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -102,7 +102,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -338,7 +338,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -355,7 +355,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd, values);
+        return (int32_t)get_integer_result(cmd, values);
     }
 
     /**
@@ -376,7 +376,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -397,7 +397,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -420,7 +420,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -441,7 +441,7 @@ public:
 
         hash_slot(key);
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
     }
 
     /**
@@ -815,7 +815,7 @@ protected:
 
         reset_hash_slot();
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
 
     }
 
@@ -860,7 +860,7 @@ protected:
 
         reset_hash_slot();
 
-        return get_integer_result(cmd);
+        return (int32_t)get_integer_result(cmd);
 
     }
 };
